PROJETS/PERSO/_ALLSEEDS/requirements.txt
ccxt
streamlit
pandas
numpy
matplotlib
scipy
pyyaml
